+++ lib/src/localization/RPLocalization.dart
@@ -1,35 +1,36 @@
 part of research_package_ui;
 
+/// Localization support. Use [translate] to translate any text used in
+/// RP or the app in general.
 class RPLocalizations {
+  Map<String, String> _localizedStrings;
+
   final Locale locale;
 
+  /// Create a localization based on [locale].
   RPLocalizations(this.locale);
 
-  // Helper method to keep the code in the widgets concise
-  // Localizations are accessed using an InheritedWidget "of" syntax
+  /// Helper method to keep the code in the widgets concise
+  /// Localizations are accessed using an [InheritedWidget] "of" syntax
   static RPLocalizations? of(BuildContext context) {
     return Localizations.of<RPLocalizations>(context, RPLocalizations);
   }
 
-  Map<String, String> _localizedStrings = {};
-
+  /// Load the language JSON file from `assets/lang/${locale.languageCode}.json`
   Future<bool> load() async {
-    // Load the language JSON file from the "lang" folder
-    String jsonString = await rootBundle.loadString('assets/lang/${locale.languageCode}.json');
+    String jsonString =
+        await rootBundle.loadString('assets/lang/${locale.languageCode}.json');
     Map<String, dynamic> jsonMap = json.decode(jsonString);
-
-    _localizedStrings = jsonMap.map((key, value) {
-      return MapEntry(key, value.toString());
-    });
+    _localizedStrings =
+        jsonMap.map((key, value) => MapEntry(key, value.toString()));
 
     return true;
   }
 
-  // This method will be called from every widget which needs a localized text
-  String? translate(String key) {
-    if (_localizedStrings[key] == null) print('Translation of $key not found');
-    return _localizedStrings[key];
-  }
+  /// Translate [text] to this [locale].
+  /// If [text] is not translated, [text] is returned untranslated.
+  String translate(String text) =>
+      (_localizedStrings.containsKey(text)) ? _localizedStrings[text] : text;
 
   // Static member to have a simple access to the delegate from the MaterialApp
   static const LocalizationsDelegate<RPLocalizations> delegate = _RPLocalizationsDelegate();
@@ -48,8 +49,7 @@ class _RPLocalizationsDelegate extends LocalizationsDelegate<RPLocalizations> {
 
   @override
   Future<RPLocalizations> load(Locale locale) async {
-    // RPLocalizations class is where the JSON loading actually runs
-    RPLocalizations localizations = new RPLocalizations(locale);
+    RPLocalizations localizations = RPLocalizations(locale);
     await localizations.load();
     return localizations;
   }