+++ lib/src/model/consent/RPDataTypeSection.dart
@@ -1,8 +1,13 @@
 part of research_package_model;
 
+/// Class representing a specific data type being collected to be shown
+/// as part of informed consent.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
-class RPDataTypeSection /* extends Serializable */ {
+class RPDataTypeSection extends Serializable {
+  /// The name of the data being collected.
   String dataName;
+
+  /// Detailed information on the type of data being collected.
   String dataInformation;
 
   RPDataTypeSection(this.dataName, this.dataInformation);