+++ CHANGELOG.md
@@ -1,5 +1,11 @@
-## 0.5.0
-- small update to carp_core
+## 0.5.3+1
+- small updates to documentation
+- making `RPTask` serializable instead of abtstract
+- fix to `translate` method
+
+## 0.5.2
+- update of json serialization in informed consent domain model
+- updated example and unit test on `RPConsentSection` for passive data collection
 
 ## 0.5.0